+++ libs/design/src/atoms/form/native-select/select/select-theme.scss
@@ -2,7 +2,7 @@
 @use '../../../../../scss/core';
 @use '../../../../../scss/theming';
 
-@mixin daff-select-theme($theme) {
+@mixin daff-native-select-theme($theme) {
 	$gray: core.daff-map-deep-get($theme, 'core.gray');
 	$base-contrast: core.daff-map-deep-get($theme, 'core.base-contrast');
 	$black: core.daff-map-deep-get($theme, 'core.black');
