+++ libs/design/src/atoms/form/native-select/select/select.component.scss
