+++ libs/design/src/atoms/form/quantity-field/quantity-field.component.spec.ts
@@ -12,7 +12,7 @@ import {
 import { By } from '@angular/platform-browser';
 
 import { DaffInputModule } from '../input/public_api';
-import { DaffNativeSelectModule } from '../select/public_api';
+import { DaffNativeSelectModule } from '../native-select/public_api';
 import { DaffQuantityFieldComponent } from './quantity-field.component';
 import { DaffQuantityInputComponent } from './quantity-input/quantity-input.component';
 import { DaffQuantitySelectComponent } from './quantity-select/quantity-select.component';
