+++ libs/design/src/atoms/form/quantity-field/quantity-field.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 
 import { DaffInputModule } from '../input/public_api';
-import { DaffNativeSelectModule } from '../select/public_api';
+import { DaffNativeSelectModule } from '../native-select/public_api';
 import { DaffQuantityFieldComponent } from './quantity-field.component';
 import { DaffQuantityInputComponent } from './quantity-input/quantity-input.component';
 import { DaffQuantitySelectComponent } from './quantity-select/quantity-select.component';
