+++ libs/design/src/public_api.ts
@@ -9,7 +9,7 @@ export * from './atoms/form/form-field/public_api';
 export * from './atoms/form/error-message/public_api';
 export * from './atoms/image/public_api';
 export * from './atoms/form/input/public_api';
-export * from './atoms/form/select/public_api';
+export * from './atoms/form/native-select/public_api';
 export * from './atoms/form/checkbox/public_api';
 export * from './atoms/form/quantity-field/public_api';
 export * from './atoms/container/public_api';
